Moved remaining schemas to folder
